requirements.txt
django
django-codemirror-widget
